+++ tax/totals_calculator_test.go
@@ -12,7 +12,6 @@ import (
 	"github.com/invopop/gobl/l10n"
 	"github.com/invopop/gobl/num"
 	"github.com/invopop/gobl/regimes/es"
-	"github.com/invopop/gobl/regimes/gr"
 	"github.com/invopop/gobl/regimes/it"
 	"github.com/invopop/gobl/regimes/pt"
 	"github.com/invopop/gobl/tax"
@@ -1229,18 +1228,12 @@ func TestTotalBySumCalculate(t *testing.T) {
 						Retained: false,
 						Rates: []*tax.RateTotal{
 							{
-								Ext: tax.Extensions{
-									gr.ExtKeyMyDATAVATCat: "1",
-								},
 								Key:     tax.RateStandard,
 								Base:    num.MakeAmount(942, 2),
 								Percent: num.NewPercentage(24, 2),
 								Amount:  num.MakeAmount(226, 2),
 							},
 							{
-								Ext: tax.Extensions{
-									gr.ExtKeyMyDATAVATCat: "2",
-								},
 								Key:     tax.RateReduced,
 								Base:    num.MakeAmount(942, 2),
 								Percent: num.NewPercentage(13, 2),