+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Finally, the `draft` flag has been removed from the header, and much more emphas
 - `es`: moved FacturaE and TicketBAI extensions to the `addons/es/facturae` and `addons/es/tbai` packages respecitvely.
 - `pt`: moved SAF-T specific extensions to `addons/pt/saft`.
 - `it`: moved SDI and FatturaPA extensions to `addons/it/sdi` with key `it-sdi-v1`.
+- `gr`: moved MyDATA to `addons/gr/mydata`, key `gr-mydata-v1`.
 
 ### Added